+++ lib/models/apphud_models/apphud_non_renewing_purchase.dart
@@ -9,17 +9,11 @@ class ApphudNonRenewingPurchase {
   final double? canceledAt;
   final bool isActive;
 
-  // TODO: remove both, not present in Android
-  final bool? isSandbox;
-  final bool? isLocal;
-
   ApphudNonRenewingPurchase({
     required this.productId,
     required this.purchasedAt,
     required this.isActive,
     this.canceledAt,
-    this.isLocal,
-    this.isSandbox,
   });
 
   factory ApphudNonRenewingPurchase.fromJson(Map<dynamic, dynamic> map) =>
@@ -29,6 +23,6 @@ class ApphudNonRenewingPurchase {
 
   @override
   String toString() {
-    return 'ApphudNonRenewingPurchase{productId: $productId, purchasedAt: $purchasedAt, canceledAt: $canceledAt, isActive: $isActive, isSandbox: $isSandbox, isLocal: $isLocal}';
+    return 'ApphudNonRenewingPurchase{productId: $productId, purchasedAt: $purchasedAt, canceledAt: $canceledAt, isActive: $isActive}';
   }
 }